+++ CHANGELOG.md
@@ -39,6 +39,9 @@
 * Add bindings for `FocusOptions`.
   [#3996](https://github.com/rustwasm/wasm-bindgen/pull/3996)
 
+* Add bindings for `RTCRtpReceiver.jitterBufferTarget`.
+  [#3968](https://github.com/rustwasm/wasm-bindgen/pull/3968)
+
 ### Changed
 
 * Stabilize Web Share API.